+++ src/main/java/me/despical/oitc/utils/ExceptionLogHandler.java
@@ -41,10 +41,12 @@ public class ExceptionLogHandler extends Handler {
 	}
 
 	@Override
-	public void close() throws SecurityException {}
+	public void close() throws SecurityException {
+	}
 
 	@Override
-	public void flush() {}
+	public void flush() {
+	}
 
 	@Override
 	public void publish(LogRecord record) {
@@ -91,7 +93,7 @@ public class ExceptionLogHandler extends Handler {
 		plugin.getLogger().log(Level.WARNING, stacktrace.toString());
 		plugin.getLogger().log(Level.WARNING, "[Reporter service] <<------------------------------[END]------------------------------>>");
 
-		record.setMessage("[OITC] We have found a bug in the code. Contact us at our official discord server (Invite link: https://discordapp.com/invite/Vhyy4HA) with the following error given above!");
+		record.setMessage("[OITC] We have found a bug in the code. Contact us at our official Discord server (Invite link: https://discordapp.com/invite/Vhyy4HA) with the following error given above!");
 	}
 
 	private boolean containsBlacklistedClass(Throwable throwable) {